+++ bot/src/main/java/edu/java/BotApplication.java
@@ -1,6 +1,6 @@
-package edu.java.bot;
+package edu.java;
 
-import edu.java.bot.configuration.ApplicationConfig;
+import edu.java.configuration.ApplicationConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
