+++ bot/src/main/java/edu/java/configuration/ApplicationConfig.java
@@ -1,4 +1,4 @@
-package edu.java.bot.configuration;
+package edu.java.configuration;
 
 import com.pengrad.telegrambot.TelegramBot;
 import edu.java.bot.exception.BotExceptionHandler;
